+++ packages/mobile/src/services/eagerLoadUtils.ts
@@ -3,7 +3,7 @@
  * protocol before libs has initialized.
  */
 
-import { getEagerDiscprov } from '@audius/common'
+import { getEagerDiscprov, makeEagerRequest } from '@audius/common'
 
 import { env } from './env'
 import { audiusLibs, waitForLibsInit } from './libs'
@@ -36,10 +36,12 @@ export const withEagerOption = async (
   } else {
     try {
       const req = eager(...args)
-      const res = await makeRequest(
+      const res = await makeEagerRequest(
         req,
         disprovEndpoint as string,
-        requiresUser
+        requiresUser,
+        localStorage,
+        env
       )
       return res
     } catch (e) {
@@ -47,81 +49,4 @@ export const withEagerOption = async (
       return normal(audiusLibs)(...args)
     }
   }
-}
-
-/**
- * Convertsa a provided URL params object to a query string
- */
-const parmsToQS = (
-  params: { [key: string]: string | string[] },
-  formatWithoutArray = false
-) => {
-  if (!params) return ''
-  return Object.keys(params)
-    .map((k) => {
-      if (Array.isArray(params[k])) {
-        return (params[k] as string[])
-          .map((val: string) =>
-            formatWithoutArray
-              ? `${encodeURIComponent(k)}=${encodeURIComponent(val)}`
-              : `${encodeURIComponent(k)}[]=${encodeURIComponent(val)}`
-          )
-          .join('&')
-      }
-      return (
-        encodeURIComponent(k) + '=' + encodeURIComponent(params[k] as string)
-      )
-    })
-    .join('&')
-}
-
-/**
- * Takes a request object provided from the audius libs API and makes the request
- * using the fetch API.
- */
-const makeRequest = async (
-  req: any,
-  endpoint: string,
-  requiresUser = false
-) => {
-  const eagerDiscprov = await getEagerDiscprov(localStorage, env)
-  const discprovEndpoint = endpoint ?? eagerDiscprov
-  const user = await localStorage.getAudiusAccountUser()
-  if (!user && requiresUser) throw new Error('User required to continue')
-
-  const headers: { [key: string]: string } = {}
-  if (user && user.user_id) {
-    headers['X-User-ID'] = user.user_id
-  }
-
-  let baseUrl = `${discprovEndpoint}/${req.endpoint}`
-  if (req.urlParams) {
-    baseUrl = `${baseUrl}${req.urlParams}`
-  }
-
-  let res: any
-  if (req?.method?.toLowerCase() === 'post') {
-    headers['Content-Type'] = 'application/json'
-    const url = `${baseUrl}?${parmsToQS(
-      req.queryParams,
-      discprovEndpoint === eagerDiscprov
-    )}`
-    res = await fetch(url, {
-      method: 'POST',
-      headers,
-      body: JSON.stringify(req.data)
-    })
-  } else {
-    const url = `${baseUrl}?${parmsToQS(
-      req.queryParams,
-      discprovEndpoint === eagerDiscprov
-    )}`
-    res = await fetch(url, {
-      headers
-    })
-  }
-
-  const json = await res.json()
-  if (json.data) return json.data
-  return json
 }
